cgi, draw route: fail early on expired sessions
Formerly, even if a session was expired, bbbike.cgi tried to draw the
image or pdf nevertheless. This resulted only in a complicated flow
for drawing the error message into the image/pdf
("empty_image_error"). And this did not work in the pdf case at all
(since PDFCairo.pm? since ever?).

Now the expired session is detected early, and a simple html error
page is returned.