radroute.html: fixed error in certain user click paths
The following click path caused an error: on the radroute.html page,
click on "routelist" button first, then use browser's back button,
then click on the "map" button of the same cycle route. This caused a
strange download dialog for a "image/" file (yes, without a real
subtype), and warnings about the use of "images/flag2_bl." and
"images/flag_ziel." (that is, without suffixes) in the errorlog.

Problem was: a click on routelist manipulated the form by basically
setting the "imagetype" form parameter to the empty string, and this
was never corrected back.