BBBikeOsmUtil: boring line/area items are drawn very faintly
Items without any "interesting" tags are drawn with a very faint
dashed line. Additionally the tooltip of such lines (but not areas) is
set to "<empty>" (previously it was the empty string).

Rationale: there are some ways in Berlin (for example the borders of
the Urstromtal) (and probably also elsewhere) which are probably just
part of relations, but have no information (i.e. tags) of their own.
Previously such ways were drawn in the default style, that is, a solid
and relatively thick line, which was too prominent.