cgi.t: use tables for output_as and imagetype tests
The file_cache functionality is only available to some of
the formats here. Use tables and and only block for
executing and testing the file cache functionality. This
probably also fixes test counts if special parameters like
-skip-mapserver are used (hopefully)

Additionally don't test file_cache functionality for geojson,
as it's not active for geojson.