Perl/Tk: plotting of opposite cycleways
Now the opposite direction is marked with a double-arrow, just like in
my maps for Garmin devices. This works only for cycleways outside of
Berlin (resp. in edit mode also inside of Berlin, due to radwege not
yet changed to use RW;RW syntax).

Rendering is also not quite optimal: the double-arrow symbol should be
on a higher rendering layer, and on zooming a redraw should be done.