moved draw_bridge and draw_tunnel_entrance to new BBBikeCanvasUtil.pm
The movement also required to add $canvas as an additional argument to
both functions. The movement also makes it now possible to test
ext/BBBikeXS/t/base.t with these two functions (previously the drawing
was just skipped). Last but not least, this fixed also a long-standing
bug where bridge and tunnel features for the overview map were drawn
on the main canvas map (the dirt around Ploetzensee...)