workaround for "make: don't know how to make w. Stop"
This problem happens if the outer make is GNU make, and there's at
least one recursive make invocation. In this case GNU make sets the
MAKEFLAGS environment variable to "w". But some test scripts may call
freebsd-make or pmake, and the content of the MAKEFLAGS variable is
interpreted as the (non-existing) target w.

Fix is to unset the MAKEFLAGS variable before calling freebsd-make,
pmake, or BSD's native make.