GPS::GpsmanData: as_gpx does not emit useless <cmt> elements
If a comment field in a gpsman file is detected as a date-time value,
then only the <time> element is emitted, but not anymore the
equivalent <cmt> element. The old behavior can be restored by setting
the autoskipcmt option to a false value.

Adapted the gps-gpsmandata.t test. Also, this test now tests with "is"
instead of "like" in some places, using a better XPath query with an
explicit index.