bbbike-snapshot.cgi: again enable snapshots from local directory
If local=1 is set, then a snapshot from the local directory is done,
like it was before introducing the default github archive link.

The code flow is somewhat inefficient (Botchecker_BBBike) being called
twice, but the local=1 URL is not announced anyway.

To avoid redefinition warnings, the "do" for Botchecker_BBBike was
replaced with a "require."