typ file: changed ids of city_limit signs
It seems that the used ids (0x72/0x01 and 0x72/0x03) were not visible
on a real Garmin device). So the ids were changed to 0x74/0x01 and
0x74/0x02.

This is also the first version of the TYP and TXT files produced by
TYPViewer. The previously used online editor at ati.land.cz is not
available anymore.