choose_from_additional_layer: don't sort gps files
Keep gps files (through heuristic detection) unsorted, in the hope
that they are already chronologically sorted (which is unfortunately
not the case for gpsman wpt files).

This has also the side effect that the search box and the A..Z bar is
not displayed anymore. In the case of unsorted data both were not
useful anyway.