gpsman <-> gpx: better roundtrip for waypoint files
As gpsman files cannot have extra attributes in the !W line, so the
TrackAttrs element have to be stored in special directives looking
like

    %!key=val

This is used to store the srt:device (for creator) and
srt:garmin_userdef_symbols_set attributes.

srt:garmin_userdef_symbols_set is needed to distinguish between the
two icon sets bike2008 and bike2014, as the gpsman file stores only
user:\d{4}. The functions in GPS::GpsmanData::GarminGPX have now to
provide or deal with the icon set.

With these changes a quite good roundtrip gpx -> gpsman -> gpx is
possible, only using uninteresting xmlns attributes and the <metadata>
element.