Strassen::GPX: -meta vs -name/-number
Make sure that <name> and <number> are created only once in presence
of a given or calculated -meta object.

This used to fail for tracks where <name> was automatically created
from start and goal point names, and <name> was additionally taken
from the title global directive. The generated xml file was invalid
according to the gpx schema.