cgi: add leaflet also to map select box
This is duplicating the plain link in the "Route as" box, but
currently it fits better in the select box, next to mapserver and
googlemaps.

The old inactive show_map() javascript function was revived to handle
leaflet as a special case: no form is posted, but the map is directly
called with window.open().

Also grouped the map select box with "static map" and "interactive
map" optgroups.