leaflet: route and feature list div
It's possible now to display a <div> with a feature or route list to
the left of the map.

The feature list is enabled in static_leaflethtml.pl with the
--show-feature-list switch. The route list is currently only enabled
in development mode (with the new CGI param devel).

The old route list implementation (displaying the route list in a
leaflet popup) is gone. The html generation code was re-used, but now
arrows instead of the lengthy direction string is displayed.

static_leaflethtml.pl has also a new --leafletver option.