Strassen: implement Strict option, and add syntax_check_* functions
Previously only $Strassen::STRICT was working, but the Strict option
had an incomplete implementation. Now the Strict option is working for
the streaming variants of Strassen. This is also documented.
Additionally, as a special case, there are new syntax_check_* class
methods which use the Strict option for a parse check of bbd file or
data string.

strassen.t tests are adapted.