don't bundle GPS::Garmin in windist anymore
A clean install of GPS::Garmin on Windows is not possible anymore,
as computers usually don't have a serial port anymore, so
Win32::SerialPort, a prereq of GPS::Garmin, does not test anymore.
Actually, gpsbabel is still bundled with bbbike and is capable of
dealing with COM* and USB trasnfers to old garmins.