BBBikeWinUtil::adjust_path
A new module BBBikeWinUtil.pm which contains a function adjust_path().
The functionality of adjust_path was always called in the bbbike
program, but not in bbbike_chooser.pl. Failing to do this could lead
to DLL errors (libxml2-2_.dll not found), unless Strawberry Perl was
already installed on this machine.