toplevel_checker: don't call ->geometry if nothing changed
Calling geometry with the same coordinates seems to actually
reposition the window in some situations. This was seen on a
Xquartz system when calling show_info. At the same time the
event system thought that the window was still at the old place,
leaving a completely broken application.

Now old and new geometry is compared and nothing is called if it
doesn't have to change.

Additionally there are a couple of new DEBUG_... directives.