docker-bbbike gui improvements for MacOSX
Start the docker-machine if it's not yet running.

The docker-machine ssh connection for the port forwarding hack is now
done automatically.

New option --docker-machine, if the wished name is not "default".

The socat and ssh processes are killed automatically using the new
scope cleanup feature of Doit.pm.

Newest Doit.pm was imported for Doit::ScopeCleanups and
setenv/unsetenv.