bbbike data download: change If-Modified-Since handling
Previously, a 304 was only sent if the If-Modified-Since header
*exactly* matched the mtime of the local file. But usually the local
file had a newer timestamp, so almost always 200 was needlessly sent.
This applied to all three data download handlers.

On the live server it was not that bad, as only a few data files
(strassen + landstrassen*) were activated in the special download
handler.