debian package: specify recommends
XML::LibXML and LWP::UserAgent are required for some bbbike
functionality, so specify it as "recommends".

Also make sure that the docker-bbbike tasks can handle
the default behavior for recommends, which is to install
these packages. This requires a new helper script install-deb
which can handle .deb installs for newer (using apt) and
older (using dpkg+apt-get hack) debians.