adapt bbox in geocoder-addr.t
Problem: the address Dudenstraße 10 resolves to (at least) two nodes,
one in the front building and the other in the rear building. The rear
building node won, and was outside the expected bbox. For now, the
bbox was adjusted, but the front building should have precedence.

Also added more diagnostics for the existing TODO test.