new option allmaps_cb for collecting links for "All Maps"
For maps which do not define callback_3_std, but the generic
callback_3 (i.e. for generating a popup menu) there's now the
possibility to define another callback just for the "All Maps" link
collection.

Use this option for OpenStreetMap and Mapillary.

Also, the list of maps in "All Maps" is now sorted in a
case-insensitive way.