bvg_checker.pl has a new -log option
The old log-writing shell command was broken (bvg_checks.log was
always empty). Also, an error was not that easy to spot, so "OK" and
failure messages are now colored. To not get ANSI sequences into the
log file a -log option was invented and this and STDERR written
separately.