get_polygon_center variants
A new variant "..._forceinside" is available in VectorUtil.pm
which was already used in the main bbbike script.

A variant "..._medialaxis" is prepared, but currently nowhere
used and still in evaluation.

get_polygon_center_best will choose the best implementation
depending on what is installed.

Adapted and extended related tests.